refactor(core): extract forward message utilities to shared koishi utils
Extract pickForwardMessageId, isForwardMessageElement, and normalizeForwardMessageId functions from read_chat_message.ts to koishi.ts utils module. This allows these utilities to be reused across different middleware and modules.

The functions are now exported from the shared utils module and imported where needed.
