+++ packages/core/src/utils/koishi.ts
@@ -145,3 +145,40 @@ export function transformToMarkdown(
     const result = render(marked.lexer(source), platform)
     return result
 }
+
+export function pickForwardMessageId(element: h): string | null {
+    const attrs = (element.attrs ?? {}) as Record<string, unknown>
+
+    // Only use the common message id field used across other message APIs.
+    // Keep both snake_case and camelCase for adapter compatibility.
+    for (const key of ['message_id', 'messageId']) {
+        const normalizedId = normalizeForwardMessageId(attrs[key])
+        if (normalizedId) return normalizedId
+    }
+
+    return null
+}
+
+export function isForwardMessageElement(element: h): boolean {
+    if (!element) return false
+    if (element.type === 'forward') return true
+    if (element.type !== 'message') return false
+
+    const attrs = element.attrs ?? {}
+    if (['true', '1'].includes(String(attrs['forward']))) return true
+
+    return false
+}
+
+export function normalizeForwardMessageId(value: unknown): string | null {
+    if (typeof value === 'number' && Number.isFinite(value)) {
+        return String(value)
+    }
+
+    if (typeof value !== 'string') {
+        return null
+    }
+
+    const trimmed = value.trim()
+    return trimmed.length > 0 ? trimmed : null
+}
